requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
boto3==1.17.72
botocore==1.20.72
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
django-cleanup==5.2.0
django-crispy-forms==1.11.2
django-storages==1.11.1
djangorestframework==3.12.4
drf-nested-routers==0.93.3
et-xmlfile==1.1.0
future==0.18.2
grpcio==1.37.1
gunicorn==20.1.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
install==1.3.4
jmespath==0.10.0
lml==0.1.0
msgpack==1.0.2
openpyxl==3.0.7
packaging==20.9
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.15.8
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.5.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
django-heroku==0.3.1
